--- thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/detection/spec/DurationAnomalyFilterSpec.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (C) 2014-2018 LinkedIn Corp. (pinot-core@linkedin.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.linkedin.thirdeye.detection.spec;
-
-public class DurationAnomalyFilterSpec extends AbstractSpec {
-  private String minDuration;
-  private String maxDuration;
-
-  public String getMinDuration() {
-    return minDuration;
-  }
-
-  public void setMinDuration(String minDuration) {
-    this.minDuration = minDuration;
-  }
-
-  public String getMaxDuration() {
-    return maxDuration;
-  }
-
-  public void setMaxDuration(String maxDuration) {
-    this.maxDuration = maxDuration;
-  }
-}
